Changed printf to use %zu for size_t (Thanks Rick Scott)

+++ src/gcalendar.c
@@ -360,8 +360,8 @@ static void gc_get_changes_calendar(OSyncObjTypeSink *sink,
 		osync_trace(TRACE_INTERNAL, "gcalendar: no changes...\n");
 		goto exit;
 	} else {
-		osync_trace(TRACE_INTERNAL, "gcalendar: changes count: %lu\n",
-			    (unsigned long)all_events.length);
+		osync_trace(TRACE_INTERNAL, "gcalendar: changes count: %zu\n",
+			    all_events.length);
 	}
 
 	// Calendar returns most recently updated event as first element
@@ -565,8 +565,8 @@ static void gc_get_changes_contact(OSyncObjTypeSink *sink,
 		osync_trace(TRACE_INTERNAL, "gcontact: no changes...\n");
 		goto exit;
 	} else
-		osync_trace(TRACE_INTERNAL, "gcontact: changes count: %lu\n",
-			    (unsigned long)all_contacts.length);
+		osync_trace(TRACE_INTERNAL, "gcontact: changes count: %zu\n",
+			    all_contacts.length);
 
 	// Contacts returns most recently updated entry as last element
 	for (i = 0; i < all_contacts.length; ++i) {
